refactor(direct-control): extract operation primitive types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves shared operation family/target/input, action approval, and operation validation result public types into src/play/operations/types.ts while keeping public facade type re-exports in index.ts.

Leaves operation request result, postcondition payloads, production choice, diplomacy/narrative closeout, public procedure schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- unit-operation.test.ts population-placement.test.ts production-choice.test.ts unit-target-action.test.ts diplomacy-response.test.ts narrative-choice.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.play.operation-wrappers.test.ts game.play.population-placement.test.ts game.play.production.test.ts game.play.unit-target.test.ts game.play.diplomacy-response.test.ts game.play.narrative.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle scans for operation primitive definitions and facade executable imports

Local package/type relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
