+++ packages/civ7-direct-control/src/play/operations/diplomacy-postconditions.ts
@@ -1,11 +1,11 @@
+import type { Civ7OperationValidationResult } from "./types.js";
+import type { Civ7ComponentId } from "../../civ7-component-id.js";
+import type { Civ7RuntimeProbe } from "../../runtime/probe.js";
 import type {
-  Civ7ComponentId,
   Civ7DiplomacyResponseInput,
   Civ7DiplomacyResponsePostcondition,
   Civ7DiplomacyResponsePostconditionClassification,
   Civ7DirectControlOptions,
-  Civ7OperationValidationResult,
-  Civ7RuntimeProbe,
 } from "../../index.js";
 import type {
   Civ7PlayNotificationSummary,
