+++ packages/civ7-direct-control/src/play/operations/narrative-postconditions.ts
@@ -1,12 +1,12 @@
+import type { Civ7OperationValidationResult } from "./types.js";
+import type { Civ7ComponentId } from "../../civ7-component-id.js";
+import type { Civ7RuntimeProbe } from "../../runtime/probe.js";
 import type {
-  Civ7ComponentId,
   Civ7DirectControlOptions,
   Civ7NarrativeChoiceCommandPayload,
   Civ7NarrativeChoiceInput,
   Civ7NarrativeChoicePostcondition,
   Civ7NarrativeChoicePostconditionClassification,
-  Civ7OperationValidationResult,
-  Civ7RuntimeProbe,
 } from "../../index.js";
 import type {
   Civ7PlayNotificationSummary,
